+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changes
 0.9.14 (unreleased)
 -------------------
 
+- Remove usage of ``Nodespaces`` behavior.
+  [rnix]
+
 - Replace deprecated use of ``Storage`` by ``MappingStorage``.
   [rnix]